make tests more complete (test localhost + all hosts in defaultHosts for all http tests) and make them (more likely) to be determinate
- there is a delay between when a http server will bind and unbind to
a host. This delay is most noticeable on node 0.12, but exists in all
versions of node. Also there is an issue with vows - it seems that batches
are being run concurrently without waiting for a previous batch to complete
-- the combination of these 2 things + the new testing of all known hosts
requires a staggering timeout to be placed around the getPort call within
the batch tests. By staggering I mean not having all timeouts expire at
the same time.

This can be proved by changing the increment() function at the top of
port-finder-test and port-finder-multiple-test to return a constant value
like 1000 or even 10000.

In hindsight, this makes some sense, except for the vows issue.